Rename vivify-shared to refresh-theme
