+++ packages/refresh-theme/graphql/fragments/dynamic-page.js
