+++ packages/refresh-theme/graphql/fragments/magazine-publication-list.js
