+++ packages/refresh-theme/graphql/fragments/magazine-publication-page.js
