+++ packages/refresh-theme/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@ac-business-media/vivify-shared",
+  "name": "@ac-business-media/refresh-theme",
   "version": "1.0.0-alpha.5",
   "author": "Jacob Bare <jacob@limit0.io>",
-  "repository": "https://github.com/base-cms-websites/ac-business-media/tree/master/packages/vivify-shared",
+  "repository": "https://github.com/base-cms-websites/ac-business-media/tree/master/packages/refresh-theme",
   "license": "MIT",
   "private": true,
   "scripts": {
