+++ packages/refresh-theme/templates/index.marko
@@ -40,7 +40,7 @@ $ const adSlots = ({ aliases }) => ({
                 name="website-optioned-content"
                 params={ sectionId: id, optionName: "Pinned", excludeContentTypes: ["Company"], limit: 6, requiresImage: true, queryFragment }
               >
-                <vivify-shared-content-hero-flow nodes=nodes />
+                <refresh-theme-content-hero-flow nodes=nodes />
               </marko-web-query>
             </div>
           </div>
