+++ sites/foodlogistics.com/package.json
@@ -20,7 +20,7 @@
     "@base-cms/object-path": "^1.0.0",
     "@base-cms/utils": "^1.0.0",
     "@base-cms/web-cli": "^1.0.2",
-    "@ac-business-media/vivify-shared": "^1.0.0-alpha.5",
+    "@ac-business-media/refresh-theme": "^1.0.0-alpha.5",
     "graphql": "^14.5.4",
     "graphql-tag": "^2.10.1"
   }
