+++ sites/foodlogistics.com/server/routes/home.js
@@ -1,6 +1,6 @@
 const { withWebsiteSection } = require('@base-cms/marko-web/middleware');
-const home = require('@ac-business-media/vivify-shared/templates/index');
-const queryFragment = require('@ac-business-media/vivify-shared/graphql/fragments/website-section-page');
+const home = require('@ac-business-media/refresh-theme/templates/index');
+const queryFragment = require('@ac-business-media/refresh-theme/graphql/fragments/website-section-page');
 
 module.exports = (app) => {
   app.get('/', withWebsiteSection({
