+++ sites/foodlogistics.com/server/styles/index.scss
@@ -29,4 +29,4 @@ $website-section-colors: (
   "3pl-4pl": #cea14c,
 );
 
-@import "../../node_modules/@ac-business-media/vivify-shared/scss/theme";
+@import "../../node_modules/@ac-business-media/refresh-theme/scss/theme";
